+++ Src/AutoSpectre.SourceGeneration.Tests/ChoicesTests.cs
@@ -1,5 +1,4 @@
-﻿using FluentAssertions;
-using Xunit;
+﻿using Xunit;
 
 namespace AutoSpectre.SourceGeneration.Tests;