+++ Src/AutoSpectre.SourceGeneration.Tests/Evaluation/ConfirmedCancelResultTests.cs
@@ -1,9 +1,4 @@
-using System;
-using System.Linq;
-using System.Reflection;
 using System.Text;
-using AutoSpectre;
-using AutoSpectre.SourceGeneration;
 using AutoSpectre.SourceGeneration.BuildContexts;
 using AutoSpectre.SourceGeneration.Evaluation;
 using FluentAssertions;