+++ Src/AutoSpectre.SourceGeneration/BuildContexts/ReuseExistingAutoSpectreFactoryPromptBuildContext.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Net.Mime;
 using System.Text;
 using AutoSpectre.SourceGeneration.Evaluation;
 using AutoSpectre.SourceGeneration.Extensions;