+++ Src/AutoSpectre.SourceGeneration/BuildContexts/TaskStepBuildContext.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Diagnostics;
 using System.Text;
 using AutoSpectre.SourceGeneration.Extensions;