+++ Src/AutoSpectre.SourceGeneration/Extensions/Specification/PropertySpecification.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
 
 namespace AutoSpectre.SourceGeneration.Extensions.Specification;