+++ Src/AutoSpectre.SourceGeneration/PromptBuildContext.cs
@@ -1,13 +1,8 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using System.Composition.Hosting.Core;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using AutoSpectre.SourceGeneration.BuildContexts;
 using AutoSpectre.SourceGeneration.Evaluation;
 using AutoSpectre.SourceGeneration.Extensions;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
 
 namespace AutoSpectre.SourceGeneration;