+++ Src/Test2/Program.cs
@@ -1,15 +1,17 @@
 ﻿using AutoSpectre;
-using Spectre.Console;
 
 
 var someForm = new SomeForm();
+someForm.Prompt();
+int i = 0;
+
 
 
 [AutoSpectreForm]
 public partial class SomeForm
 {
     [SelectPrompt(Title = "Enter your name")]
-    public partial string[] Name {get;set;}
+    public partial string[] Name { get; set; } = [];
     
     public string[] NameSource => ["Jeppe", "Peter","Ulrik"];
     public List<string> NameDefault = ["Jeppe","Peter"];